Cleaned up the Qt code
