+++ include/ibc/qt/image_scroll_area.h
@@ -55,13 +55,98 @@ namespace ibc
 
   public:
     // Constructors and Destructor ---------------------------------------------
-    ImageScrollArea(ImageView *image, QWidget *parent = Q_NULLPTR);
-    virtual ~ImageScrollArea();
+    // -------------------------------------------------------------------------
+    // ImageScrollArea
+    // -------------------------------------------------------------------------
+    ImageScrollArea(ImageView *image, QWidget *parent = Q_NULLPTR)
+      : QScrollArea(parent)
+    {
+      mImageView = image;
+
+      setBackgroundRole(QPalette::Dark);
+      setAlignment(Qt::AlignHCenter | Qt::AlignVCenter);
+      setWidget(mImageView);
+    }
+    // -------------------------------------------------------------------------
+    // ~ImageScrollArea
+    // -------------------------------------------------------------------------
+    virtual ~ImageScrollArea()
+    {
+    }
 
   protected:
-    void wheelEvent(QWheelEvent *wEvent) override;
-
+    // Member variables --------------------------------------------------------
     ImageView *mImageView;
+    QPoint  mMousePreviousPos;
+
+    // Member functions --------------------------------------------------------
+    // -------------------------------------------------------------------------
+    // mousePressEvent
+    // -------------------------------------------------------------------------
+    virtual void  mousePressEvent(QMouseEvent *event)
+    {
+      if (event->button() == Qt::LeftButton)
+      {
+        mMousePreviousPos = event->pos();
+      }
+    }
+    // -------------------------------------------------------------------------
+    // mouseMoveEvent
+    // -------------------------------------------------------------------------
+    virtual void  mouseMoveEvent(QMouseEvent *event)
+    {
+      QPoint  diff = mMousePreviousPos - event->pos();
+      setScrollBarValueDiff(horizontalScrollBar(), diff.x());
+      setScrollBarValueDiff(verticalScrollBar(), diff.y());
+      mMousePreviousPos = event->pos();
+    }
+    // -------------------------------------------------------------------------
+    // mouseReleaseEvent
+    // -------------------------------------------------------------------------
+    virtual void  mouseReleaseEvent(QMouseEvent *event)
+    {
+    }
+    // -------------------------------------------------------------------------
+    // wheelEvent
+    // -------------------------------------------------------------------------
+    void wheelEvent(QWheelEvent *wEvent) override
+    {
+      QPoint delta = wEvent->angleDelta();
+      QPoint pos = mImageView->mapFromGlobal(wEvent->globalPos());
+      QSize size = mImageView->size();
+      double hOffset = pos.x() * (1.0 / mImageView->getZoomScale()); // Offset from the origin
+      double vOffset = pos.y() * (1.0 / mImageView->getZoomScale());
+      int x_offset = pos.x() - horizontalScrollBar()->value();  // Offset on the display
+      int y_offset = pos.y() - verticalScrollBar()->value();
+
+      int step = delta.y() / ImageView::MOUSE_WHEEL_ZOOM_STEP;
+      double scale = mImageView->calcZoomScale(step);
+      mImageView->setZoomScale(scale);
+
+      if (pos.x() >= 0 && pos.x() < size.width() &&
+        pos.y() >= 0 && pos.y() < size.height())
+      {
+        int h = (int)(hOffset * mImageView->getZoomScale()) - x_offset;
+        int v = (int)(vOffset * mImageView->getZoomScale()) - y_offset;
+        setScrollBarValue(horizontalScrollBar(), h);
+        setScrollBarValue(verticalScrollBar(), v);
+      }
+    }
+
+  private:
+    // Member functions --------------------------------------------------------
+    void  setScrollBarValue(QScrollBar *inBar, int inNewValue)
+    {
+      if (inNewValue < inBar->minimum())
+        inNewValue = inBar->minimum();
+      if (inNewValue > inBar->maximum())
+        inNewValue = inBar->maximum();
+      inBar->setValue(inNewValue);
+    }
+    void  setScrollBarValueDiff(QScrollBar *inBar, int inDiff)
+    {
+      setScrollBarValue(inBar, inBar->value() + inDiff);
+    }
   };
  };
 };
